prime factor

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-long int count, i, j, prime, largest;
+int count, i, j, prime, largest;
 
 for (i = 2; i <= 612852475143; i++)
 {
@@ -33,6 +33,6 @@ largest = i;
 }
 
 }
-printf("%lu\n", largest);
+printf("%d\n", largest);
 return (0);
 }
